add validation to not duplicate employee

+++ app.js
@@ -91,9 +91,13 @@ function getData(event) {
     let level = document.getElementById("level").value;
     let imageUrl = document.getElementById("imageUrl").value;
     let employeeForm = new employee(id, fullName, department, level, imageUrl);
-    localStorage.setItem("employee" + event.timeStamp, JSON.stringify(employeeForm));
-    console.log(employeeForm);
-    createCard(employeeForm);
+    if (checkEmployeeIfExist(id)) {
+        alert("Employee already exist")
+    } else {
+        localStorage.setItem("employee" + event.timeStamp, JSON.stringify(employeeForm));
+        console.log(employeeForm);
+        createCard(employeeForm);
+    }
     event.preventDefault();
 
 }
@@ -158,6 +162,21 @@ function getAllStorageData() {
         }
     }
     return values;
+}
+
+function checkEmployeeIfExist(employeeID) {
+    const employeeInStorage = getAllStorageData();
+
+    if (employeeInStorage.length > 0) {
+        for (let index = 0; index < employeeInStorage.length; index++) {
+            const element = employeeInStorage[index];
+            if (element.employeeID == employeeID) {
+                return true
+            }
+        }
+    }
+    return false;
+
 }
